requirements.txt
numpy==1.11.2
python-midi==0.2.4
matplotlib==1.5.3
seaborn==0.7.1
msgpack-python==0.4.8